requirements.txt
langchain==0.0.190
python-dotenv==1.0.0
streamlit==1.23.1
